% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{biplotAnyPC}
\alias{biplotAnyPC}
\title{Title}
\usage{
biplotAnyPC(
  computedPCA,
  x,
  y,
  shapeBY = NULL,
  colorBY = NULL,
  sizeBY = NULL,
  labelBY = NULL,
  genes = NULL,
  genesLabel = NULL
)
}
\arguments{
\item{genes}{}
}
\description{
Title
}
