% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadSeq-package.R
\docType{package}
\name{broadSeq-package}
\alias{broadSeq}
\alias{broadSeq-package}
\title{broadSeq : for streamlined exploration of RNA-seq data}
\description{
This package helps user to do easily RNA-seq data analysis with multiple methods (usually which needs many different input formats). Here the user will provid the expression data as a SummarizedExperiment object and will get results from different methods. It will help user to quickly evaluate different methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dasroy/broadSeq}
  \item Report bugs at \url{https://github.com/dasroy/broadSeq/issues}
}

}
\author{
\strong{Maintainer}: Rishi Das Roy \email{rishi.dasroy@gmail.com} (\href{https://orcid.org/0000-0002-3276-7279}{ORCID})

}
