% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{getFeatureLoadRanking}
\alias{getFeatureLoadRanking}
\title{Title}
\usage{
getFeatureLoadRanking(
  computedPCA,
  pcs = 1:5,
  topN = 10,
  keep = c("symbol", "gene_biotype", "seq_name", "Class")
)
}
\arguments{
\item{keep}{}
}
\description{
Title
}
