% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotAnyPC}
\alias{plotAnyPC}
\title{Title}
\usage{
plotAnyPC(
  computedPCA,
  x,
  y,
  shapeBY = NULL,
  colorBY = NULL,
  sizeBY = NULL,
  labelBY = NULL
)
}
\arguments{
\item{labelBY}{}
}
\description{
Title
}
