% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{plotHeatmapCluster}
\alias{plotHeatmapCluster}
\title{plotHeatmapCluster integrates SummarizedExperiment with pheatmap and return object as ggplot}
\usage{
plotHeatmapCluster(
  se,
  scaledAssay = "vst",
  ntop = 500L,
  genes = NULL,
  show_geneAs = NULL,
  annotation_col = NA,
  annotation_row = NA,
  ...
)
}
\arguments{
\item{annotation_col}{a character vector of colnames of colData(se)}
}
\description{
plotHeatmapCluster integrates SummarizedExperiment with pheatmap and return object as ggplot
}
