% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleAssay_plot}
\alias{sampleAssay_plot}
\title{Useful to visualize transformed/normalized assay with count assay. Plots 'boxplot' of any assay for each sample. Aesthetic can be added from colData.}
\usage{
sampleAssay_plot(smrExpt, assayName = "counts", ...)
}
\arguments{
\item{...}{}
}
\description{
Useful to visualize transformed/normalized assay with count assay. Plots 'boxplot' of any assay for each sample. Aesthetic can be added from colData.
}
