% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_baySeq}
\alias{use_baySeq}
\title{Title}
\usage{
use_baySeq(
  smrExpt,
  class_id,
  control,
  treatment,
  rank = FALSE,
  baySeq.group = "DE",
  baySeq.FDR = 1,
  ...
)
}
\arguments{
\item{baySeq.FDR}{}
}
\description{
Title
}
