% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_edgeR}
\alias{use_edgeR}
\title{Title}
\usage{
use_edgeR(
  smrExpt,
  class_id,
  control,
  treatment,
  rank = FALSE,
  edgeR.n = Inf,
  edgeR.adjust.method = "BH",
  edgeR.sort.by = "PValue",
  option = "GLM",
  ...
)
}
\arguments{
\item{edgeR.sort.by}{}
}
\description{
Title
}
