% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_limma_trend}
\alias{use_limma_trend}
\title{use_limma_trend}
\usage{
use_limma_trend(smrExpt, class_id, control, treatment, rank = FALSE, ...)
}
\arguments{
\item{smrExpt}{a \code{SummarizedExperiment} object}

\item{class_id}{one of the colnames of colData(smrExpt) and must be a factor.}

\item{control}{one of the factor values of class_id}

\item{treatment}{another factor values of class_id}

\item{rank}{(TRUE/FALSE) whether to rank genes}

\item{...}{additional arguments for limma::topTable}
}
\value{
results of limma::topTable
}
\description{
use_limma_trend
}
