% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_NOIseq}
\alias{use_NOIseq}
\title{Title}
\usage{
use_NOIseq(smrExpt, class_id, control, treatment, rank = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
We can use as(smrExpt, "ExpressionSet") to get an eSet easily but
then it will be hard to refer the treatment and control. The order
of expt_factors influence the log fold change sign. To keep it
comparable to other methods the expt_factors is used.
}
