% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Title}
\usage{
volcanoPlot(
  df,
  pValName,
  lFCName,
  sigThreshold = 0.05,
  logFCThreshold = 1,
  labelName = NULL,
  selectedLabel = NULL,
  palette = "nejm"
)
}
\arguments{
\item{palette}{one of "npg" ,"aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "nejm" or similar to viridis::cividis(3)}
}
\description{
Title
}
