% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genes_plot}
\alias{genes_plot}
\title{Title}
\usage{
genes_plot(smrExpt, features, facet.by = "feature", ...)
}
\arguments{
\item{y}{facet.by must be one of the column names of rowData(smrExpt). default "feature" which is equivalent to rownames of rowData}
}
\description{
Title
}
