% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeEdgerCPM}
\alias{normalizeEdgerCPM}
\title{Title}
\usage{
normalizeEdgerCPM(se, normLibSizes.method = "TMM", cpm.log = TRUE, ...)
}
\arguments{
\item{normLibSizes.method}{"none" ,"TMM"}

\item{...}{passed to normLibSizes function}
}
\description{
Title
}
