% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Title}
\usage{
plot_MDS(
  se,
  scaledAssay = "vst",
  ntop = 500L,
  genes = NULL,
  shapeBY = NULL,
  colorBY = NULL,
  sizeBY = NULL,
  labelBY = NULL
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
