% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{plotHeatmapCluster}
\alias{plotHeatmapCluster}
\title{Title}
\usage{
plotHeatmapCluster(
  se,
  scaledAssay = "vst",
  ntop = 500L,
  genes = NULL,
  annotation_col = c(""),
  ...
)
}
\arguments{
\item{annotation_col}{a character vector of colnames of colData(se)}
}
\description{
Title
}
