% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringVisualization.R
\name{prcompTidy}
\alias{prcompTidy}
\title{Perform Principal Components Analysis on a DESeqTransform object
\href{https://gist.github.com/tavareshugo/5ca8a5e18fedc3f23f5ec4b09a9fc906}{reused code}}
\usage{
prcompTidy(se, scaledAssay = "vst", ntop = 500L, genes = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
a list with four \code{data.frame} objects: pc_scores, eigen_values,
loadings (eigen vectors) and the original data.
}
\description{
This function is based on the \code{DESeq2::plotPCA()} function, but returns the
results of \code{prcomp} in a tidy list format. This is more flexible for further
custom plotting and exploring factor loadings of the PCA.
}
