% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_df}
\alias{round_df}
\title{Applies round function on numeric columns of a data.frame.}
\usage{
round_df(df, digits)
}
\arguments{
\item{digits}{}
}
\description{
Applies round function on numeric columns of a data.frame.
}
