% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_limma}
\alias{use_limma}
\title{Title}
\usage{
use_limma(
  smrExpt,
  class_id,
  control,
  treatment,
  rank = FALSE,
  useVoom = TRUE,
  showPlot = FALSE,
  limma.adjust = "BH",
  limma.sort.by = "p",
  limma.number = Inf,
  ...
)
}
\arguments{
\item{limma.number}{}
}
\description{
Title
}
