% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_multDE}
\alias{use_multDE}
\title{Title}
\usage{
use_multDE(
  deFun_list,
  return.df = FALSE,
  smrExpt,
  class_id,
  control,
  treatment,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
