% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{use_SAMseq}
\alias{use_SAMseq}
\title{samr package is not installable as package 'impute' is not available. Please
install this package manually before using this.}
\usage{
use_SAMseq(smrExpt, class_id, control, treatment, rank = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
samr package is not installable as package 'impute' is not available. Please
install this package manually before using this.
}
